Restructure array serialization
